feat: :sparkles: add height: fit-content to styleddefaultcomponents and move all icons to another file
